Driver data packing + improvements (#261)
* [Driver] add driver_base.py as own template file + comments

* [Driver] also move validation to won template + use in transform

* [Driver] more comments

* [Driver] suggested updates from PYNQ team + async mode exec_on_buffers

* [Driver] allow smaller batchsize in execute_on_buffers

* [Driver] optimize buffer alloc a bit

* [Driver] wait condition fix

* [Deps] update finn-base

* [Deps] update finn-base

* [Driver] enable fast_mode, expose more benchmarks

* [Deps] update finn-base
